任务66：1102-角色攻击功能的开发（完成）

+++ RedAlertProject/Assets/Scripts/CharacterSystem/AttrStrategy/EnemyAttrStrategy.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class EnemyAttrStrategy : IAttrStrategy
 {
-    public int GetCritDmg(int critRate)
+    public int GetCritDmg(float critRate)
     {
         if (UnityEngine.Random.Range(0,1f)<critRate)
         {
