+++ RedAlertProject/Assets/Scripts/CharacterSystem/AttrStrategy/IAttrStrategy.cs
@@ -7,7 +7,7 @@ public interface IAttrStrategy
 {
     int GetExtraHPValue(int lv);
     int GetDmgDescValue(int lv);
-    int GetCritDmg(int critRate);
+    int GetCritDmg(float critRate);
     
 }
